requirements.txt
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
black==23.7.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
flake8==6.1.0
Flask==3.0.0
frozenlist==1.4.0
future==0.18.3
google-auth==2.23.2
google-auth-oauthlib==1.1.0
gspread==5.11.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==3.5.0
load-dotenv==0.1.0
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.1
pandas==2.1.1
pathspec==0.11.2
platformdirs==3.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
wrapt==1.15.0
yarl==1.9.2
